Integration and testing (#2)
* Fixed url

* Tuning and tidying

* Enabled tdd debugging

* Tuning values and fixing some bugs

* Moved tuning parameters into a separate file

* Fixed precision error

* Apply suggestions from code review

* Apply suggestions from code review

+++ src/hardware/plugin/motors_plugin.c
@@ -28,7 +28,7 @@ static void update_output(motor_power_t power, reg_t* direction, uint8_t directi
     power = -power;
   }
 
-  uint16_t compare = (period * power) / motor_power_max;
+  uint16_t compare = ((uint64_t)period * power) / motor_power_max;
   *ccrh = compare >> 8;
   *ccrl = compare & 0xFF;
 }
